Rename script.js to bibleview.js
