+++ bibleview-js/src/bibleview.js
